net: imple ResolverImpl
